+++ CHANGES_NEXT_RELEASE
@@ -11,3 +11,4 @@
 * Issue #280  Reimplementation of POST /entities/attrs, using the new mongo driver (mongoc) - only used if -experimental is set
 * Issue #280  Reimplementation of DELETE /entities/attrs/{attrName}, using the new mongo driver (mongoc) - only used if -experimental is set
 * Issue #280  Better health check - before accepting, a connection to the mongo server is done (and postgres if troe is enabled)
+* Issue #280  Reimplementation of PATCH /entities/attrs/{attrName}, using the new mongo driver (mongoc) - only used if -experimental is set
