+++ src/lib/orionld/serviceRoutines/orionldPatchAttribute.h
@@ -3,7 +3,7 @@
 
 /*
 *
-* Copyright 2018 FIWARE Foundation e.V.
+* Copyright 2022 FIWARE Foundation e.V.
 *
 * This file is part of Orion-LD Context Broker.
 *
