+++ src/lib/orionld/serviceRoutines/orionldPostBatchUpdate.cpp
@@ -187,10 +187,11 @@ bool orionldPostBatchUpdate(void)
       // Alterations need the complete API entity (I might change that for the complete DB Entity ...)
       // dbModelToApiEntity is DESTRUCTIVE, so I need to clone the 'finalDbEntityP' first
       //
-      KjNode* dbEntityCopy    = kjClone(orionldState.kjsonP, finalDbEntityP);
-      KjNode* finalApiEntityP = dbModelToApiEntity(dbEntityCopy, false, entityId);
+      KjNode* dbEntityCopy      = kjClone(orionldState.kjsonP, finalDbEntityP);
+      KjNode* finalApiEntityP   = dbModelToApiEntity(dbEntityCopy, false, entityId);
+      KjNode* initialDbEntityP  = NULL;  // FIXME: initialDbEntity might not be NULL
 
-      alteration(entityId, entityType, finalApiEntityP, inEntityP);
+      alteration(entityId, entityType, finalApiEntityP, inEntityP, initialDbEntityP);
     }
 
     inEntityP = next;
