+++ test/functionalTest/cases/0000_ngsild/ngsild_langprop-patch-attribute.test
@@ -182,15 +182,15 @@ MongoDB server version: REGEX(.*)
 	],
 	"attrs" : {
 		"https://uri=etsi=org/ngsi-ld/default-context/L" : {
+			"type" : "LanguageProperty",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
 			"value" : {
 				"en" : "English",
 				"es" : "Castellano",
 				"sv" : "Swedish"
 			},
-			"type" : "LanguageProperty",
-			"mdNames" : [ ],
-			"creDate" : REGEX(.*),
-			"modDate" : REGEX(.*)
+			"mdNames" : [ ]
 		}
 	},
 	"creDate" : REGEX(.*),
