+++ test/functionalTest/cases/0000_ngsild/ngsild_new_first_notification.test
@@ -118,6 +118,13 @@ echo
 echo
 
 
+echo "06. See E1 in DB"
+echo "================"
+mongoCmd2 ftest "db.entities.findOne()"
+echo
+echo
+
+
 --REGEXPECT--
 01. Create an Entity E01 with attr P1 and P2
 ============================================
@@ -183,6 +190,52 @@ Ngsild-Attribute-Format: Normalized
     "type": "Notification"
 }
 =======================================
+
+
+06. See E1 in DB
+================
+MongoDB shell version REGEX(.*)
+connecting to: REGEX(.*)
+MongoDB server version: REGEX(.*)
+{
+	"_id" : {
+		"id" : "urn:ngsi-ld:E01",
+		"type" : "http://example.org/T",
+		"servicePath" : "/"
+	},
+	"attrNames" : [
+		"http://example.org/P1",
+		"http://example.org/P2",
+		"http://example.org/P3"
+	],
+	"attrs" : {
+		"http://example=org/P1" : {
+			"type" : "Property",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
+			"value" : 1,
+			"mdNames" : [ ]
+		},
+		"http://example=org/P2" : {
+			"type" : "Property",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
+			"value" : "ok",
+			"mdNames" : [ ]
+		},
+		"http://example=org/P3" : {
+			"type" : "Property",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
+			"value" : 3,
+			"mdNames" : [ ]
+		}
+	},
+	"creDate" : REGEX(.*),
+	"modDate" : REGEX(.*),
+	"lastCorrelator" : ""
+}
+bye
 
 
 --TEARDOWN--
