+++ test/functionalTest/cases/0000_ngsild/ngsild_new_issue_0504.test
@@ -1,0 +1,62 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+Issue 504
+
+--SHELL-INIT--
+# Copied from ngsild_issue_0504.test
+export BROKER=orionld
+dbInit CB
+brokerStart CB 0 IPv4 -experimental
+
+--SHELL--
+
+#
+# 01. Send a PATCH request with Content-Type app/json and a simple integer as payload data
+#
+
+echo "01. Send a PATCH request with Content-Type app/json and a simple integer as payload data"
+echo "========================================================================================"
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:Product:001/attrs/price/value --payload "87" -X PATCH
+echo
+echo
+
+
+--REGEXPECT--
+01. Send a PATCH request with Content-Type app/json and a simple integer as payload data
+========================================================================================
+HTTP/1.1 400 Bad Request
+Content-Length: 160
+Content-Type: application/json
+Date: REGEX(.*)
+
+{
+    "detail": "The payload data must be either a JSON Array or a JSON Object",
+    "title": "Invalid Payload",
+    "type": "https://uri.etsi.org/ngsi-ld/errors/InvalidRequest"
+}
+
+
+--TEARDOWN--
+brokerStop CB
+dbDrop CB
