+++ test/functionalTest/cases/0000_ngsild/ngsild_new_no_notification_on_nochange.test
@@ -1,0 +1,236 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+No notification on update that doesn't change anything
+
+--SHELL-INIT--
+export BROKER=orionld
+dbInit CB
+brokerStart CB 0 IPv4 -noNotifyFalseUpdate -experimental
+accumulatorStart --pretty-print
+
+--SHELL--
+
+#
+# 01. Make accumulator subscribe to changes in E1 on CB
+# 02. Create entity E1 - get notification
+# 03. Dump and reset accumulator - see one notification
+# 04. Update E1 but without changing any values
+# 05. Dump and reset accumulator - see no notification
+# 06. Update E1:P1, changing its value - to provoke a notification
+# 07. Dump and reset accumulator - see one notification
+#
+
+echo "01. Make accumulator subscribe to changes in E1 on CB"
+echo "====================================================="
+payload='{
+  "id": "urn:ngsi-ld:subscriptions:copy-e1",
+  "type": "Subscription",
+  "name": "notification on E1",
+  "entities": [
+    {
+      "id": "urn:ngsi-ld:entities:E1",
+      "type": "T1"
+    }
+  ],
+  "notification": {
+    "endpoint": {
+      "uri": "http://127.0.0.1:'${LISTENER_PORT}'/notify"
+    }
+  }
+}'
+orionCurl --url /ngsi-ld/v1/subscriptions --payload "$payload"
+echo
+echo
+
+
+echo "02. Create entity E1 - get notification"
+echo "======================================="
+payload='{
+  "id": "urn:ngsi-ld:entities:E1",
+  "type": "T1",
+  "P1": {
+    "type": "Property",
+    "value": 1
+  },
+  "R1": {
+    "type": "Relationship",
+    "object": "urn:ngsi-ld:entities:E2"
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities --payload "$payload"
+echo
+echo
+
+
+echo "03. Dump and reset accumulator - see one notification"
+echo "====================================================="
+accumulatorDump
+accumulatorReset
+echo
+echo
+
+
+echo "04. Update E1 but without changing any values"
+echo "============================================="
+payload='{
+  "value": 1
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/P1 -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "05. Dump and reset accumulator - see no notification"
+echo "===================================================="
+accumulatorDump
+accumulatorReset
+echo
+echo
+
+
+echo "06. Update E1:P1, changing its value - to provoke a notification"
+echo "================================================================"
+payload='{
+  "value": 6
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/P1 -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "07. Dump and reset accumulator - see one notification"
+echo "====================================================="
+accumulatorDump
+accumulatorReset
+echo
+echo
+
+
+--REGEXPECT--
+01. Make accumulator subscribe to changes in E1 on CB
+=====================================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/subscriptions/urn:ngsi-ld:subscriptions:copy-e1
+
+
+
+02. Create entity E1 - get notification
+=======================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1
+
+
+
+03. Dump and reset accumulator - see one notification
+=====================================================
+POST http://127.0.0.1:9997/notify?subscriptionId=urn:ngsi-ld:subscriptions:copy-e1
+Content-Length: 338
+User-Agent: orionld/REGEX(.*)
+Accept: application/json
+Content-Type: application/json
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+Ngsild-Attribute-Format: Normalized
+
+{
+    "data": [
+        {
+            "P1": {
+                "type": "Property",
+                "value": 1
+            },
+            "R1": {
+                "object": "urn:ngsi-ld:entities:E2",
+                "type": "Relationship"
+            },
+            "id": "urn:ngsi-ld:entities:E1",
+            "type": "T1"
+        }
+    ],
+    "id": "urn:ngsi-ld:Notification:REGEX(.*)",
+    "notifiedAt": "202REGEX(.*)",
+    "subscriptionId": "urn:ngsi-ld:subscriptions:copy-e1",
+    "type": "Notification"
+}
+=======================================
+
+
+04. Update E1 but without changing any values
+=============================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+05. Dump and reset accumulator - see no notification
+====================================================
+
+
+06. Update E1:P1, changing its value - to provoke a notification
+================================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+07. Dump and reset accumulator - see one notification
+=====================================================
+POST http://127.0.0.1:9997/notify?subscriptionId=urn:ngsi-ld:subscriptions:copy-e1
+Content-Length: 338
+User-Agent: orionld/REGEX(.*)
+Accept: application/json
+Content-Type: application/json
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+Ngsild-Attribute-Format: Normalized
+
+{
+    "data": [
+        {
+            "P1": {
+                "type": "Property",
+                "value": 6
+            },
+            "R1": {
+                "object": "urn:ngsi-ld:entities:E2",
+                "type": "Relationship"
+            },
+            "id": "urn:ngsi-ld:entities:E1",
+            "type": "T1"
+        }
+    ],
+    "id": "urn:ngsi-ld:Notification:REGEX(.*)",
+    "notifiedAt": "202REGEX(.*)",
+    "subscriptionId": "urn:ngsi-ld:subscriptions:copy-e1",
+    "type": "Notification"
+}
+=======================================
+
+
+--TEARDOWN--
+brokerStop CB
+accumulatorStop
+dbDrop CB
