+++ test/functionalTest/cases/0000_ngsild/ngsild_new_no_notification_on_nochange_2.test
@@ -1,0 +1,280 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+No notification if no real change
+
+--SHELL-INIT--
+# Copied from ngsild_no_notification_on_nochange_2.test
+export BROKER=orionld
+dbInit CB
+brokerStart CB 0 IPv4 -noNotifyFalseUpdate -experimental
+accumulatorStart --pretty-print 127.0.0.1 ${LISTENER_PORT}
+
+--SHELL--
+
+#
+# 01. Create a subscription matching Entity urn:ngsi-ld:entities:E1
+# 02. Create Entity urn:ngsi-ld:entities:E1
+# 03. Dump and reset accumulator and see one notification
+# 04. Update E1.number to the same value it already has
+# 05. Dump accumulator, see no notification
+# 06. Update E1.bool to the same value it already has
+# 07. Dump accumulator, see no notification
+# 08. Update E1.string to the same value it already has
+# 09. Dump accumulator, see no notification
+# 10. Update E1.hobject to the same value it already has
+# 11. Dump accumulator, see no notification
+# 12. Update E1.array to the same value it already has
+# 13. Dump accumulator, see no notification
+#
+
+echo "01. Create a subscription matching Entity urn:ngsi-ld:entities:E1"
+echo "================================================================="
+payload='{
+  "id": "urn:ngsi-ld:Subscription:01",
+  "type": "Subscription",
+  "entities": [{"id": "urn:ngsi-ld:entities:E1", "type": "T"}],
+  "notification": {
+    "format": "normalized",
+    "endpoint": {
+      "uri": "http://127.0.0.1:'${LISTENER_PORT}'/notify"
+    }
+  }
+}'
+orionCurl --url /ngsi-ld/v1/subscriptions --payload "$payload"
+echo
+echo
+
+
+echo "02. Create Entity urn:ngsi-ld:entities:E1"
+echo "========================================="
+payload='{
+  "id": "urn:ngsi-ld:entities:E1",
+  "type": "T",
+  "string": {
+    "type": "Property",   
+    "value": "Mercedes"
+  },
+  "number": {
+    "type": "Property",
+    "value": 1
+  },
+  "bool": {
+    "type": "Property",
+    "value": true
+  },
+  "hobject": {
+    "type": "Property",
+    "value": {"a": 1, "b": 2, "c": 3, "d": "4", "E": [1,2,3]}
+  },
+  "array": {
+    "type": "Property",
+    "value": [ 1, 2, { "a": 1, "b": 2 } ]
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities --payload "$payload"
+echo
+echo
+
+
+echo "03. Dump and reset accumulator and see one notification"
+echo "======================================================="
+accumulatorDump
+accumulatorReset
+echo
+echo
+
+
+echo "04. Update E1.number to the same value it already has"
+echo "====================================================="
+payload='{
+  "type": "Property",
+  "value": 1
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/number -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "05. Dump accumulator, see no notification"
+echo "========================================="
+accumulatorDump
+echo
+echo
+
+
+echo "10. Update E1.hobject to the same value it already has"
+echo "======================================================"
+payload='{
+  "type": "Property",
+  "value": {"d": "4", "E": [1,2,3], "b": 2, "a": 1, "c": 3 }
+
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/hobject -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "11. Dump accumulator, see no notification"
+echo "========================================="
+accumulatorDump
+echo
+echo
+
+
+echo "12. Update E1.array to the same value it already has"
+echo "===================================================="
+payload='{
+  "type": "Property",
+  "value": [ 1, 2, { "b": 2, "a": 1 } ]
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/array -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "13. Dump accumulator, see no notification"
+echo "========================================="
+accumulatorDump
+echo
+echo
+
+
+--REGEXPECT--
+01. Create a subscription matching Entity urn:ngsi-ld:entities:E1
+=================================================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/subscriptions/urn:ngsi-ld:Subscription:01
+
+
+
+02. Create Entity urn:ngsi-ld:entities:E1
+=========================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1
+
+
+
+03. Dump and reset accumulator and see one notification
+=======================================================
+POST http://127.0.0.1:9997/notify?subscriptionId=urn:ngsi-ld:Subscription:01
+Content-Length: 493
+User-Agent: orionld/REGEX(.*)
+Accept: application/json
+Content-Type: application/json
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+Ngsild-Attribute-Format: Normalized
+
+{
+    "data": [
+        {
+            "array": {
+                "type": "Property",
+                "value": [
+                    1,
+                    2,
+                    {
+                        "a": 1,
+                        "b": 2
+                    }
+                ]
+            },
+            "bool": {
+                "type": "Property",
+                "value": true
+            },
+            "hobject": {
+                "type": "Property",
+                "value": {
+                    "E": [
+                        1,
+                        2,
+                        3
+                    ],
+                    "a": 1,
+                    "b": 2,
+                    "c": 3,
+                    "d": "4"
+                }
+            },
+            "id": "urn:ngsi-ld:entities:E1",
+            "number": {
+                "type": "Property",
+                "value": 1
+            },
+            "string": {
+                "type": "Property",
+                "value": "Mercedes"
+            },
+            "type": "T"
+        }
+    ],
+    "id": "urn:ngsi-ld:Notification:REGEX(.*)",
+    "notifiedAt": "20REGEX(.*)",
+    "subscriptionId": "urn:ngsi-ld:Subscription:01",
+    "type": "Notification"
+}
+=======================================
+
+
+04. Update E1.number to the same value it already has
+=====================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+05. Dump accumulator, see no notification
+=========================================
+
+
+10. Update E1.hobject to the same value it already has
+======================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+11. Dump accumulator, see no notification
+=========================================
+
+
+12. Update E1.array to the same value it already has
+====================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+13. Dump accumulator, see no notification
+=========================================
+
+
+--TEARDOWN--
+brokerStop CB
+accumulatorStop
+dbDrop CB
