+++ test/functionalTest/cases/0000_ngsild/ngsild_new_patch_with_ContentType-merge-patch+json.test
@@ -1,0 +1,142 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+Patching attributes and single attribute using Content-Type "application/merge-patch+json"
+
+--SHELL-INIT--
+# Copied from ngsild_patch_with_ContentType-merge-patch+json.test
+export BROKER=orionld
+dbInit CB
+brokerStart CB 0 IPv4 -experimental
+
+--SHELL--
+
+#
+# 01. Create an entity E1 with 2 attrs: P1, R1
+# 02. PATCH E1, modifying both P1 and R1, with Content-Type application/merge-patch+json
+# 03. PATCH E1/P1, with Content-Type application/merge-patch+json
+# 04. GET E1 - make sure step 02 and 03 worked
+#
+
+echo "01. Create an entity E1 with 2 attrs: P1, R1"
+echo "============================================"
+payload='{
+  "id": "urn:ngsi-ld:entity:E1",
+  "type": "Device",
+  "P1": {
+    "type": "Property",
+    "value": 1
+  },
+  "R1": {
+    "type": "Relationship",
+    "object": "urn:ngsi-ld:entity:E1"
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities --payload "$payload"
+echo
+echo
+
+
+echo "02. PATCH E1, modifying both P1 and R1, with Content-Type application/merge-patch+json"
+echo "======================================================================================"
+payload='{
+  "P1": {
+    "type": "Property",
+    "value": 2
+  },
+  "R1": {
+    "type": "Relationship",
+    "object": "urn:ngsi-ld:entity:E2"
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entity:E1/attrs -X PATCH --payload "$payload" --in application/merge-patch+json
+echo
+echo
+
+
+echo "03. PATCH E1/P1, with Content-Type application/merge-patch+json"
+echo "==============================================================="
+payload='{
+  "value": 3
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entity:E1/attrs/P1 -X PATCH --payload "$payload" --in application/merge-patch+json
+echo
+echo
+
+
+echo "04. GET E1 - make sure step 02 and 03 worked"
+echo "============================================"
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entity:E1
+echo
+echo
+
+
+--REGEXPECT--
+01. Create an entity E1 with 2 attrs: P1, R1
+============================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/entities/urn:ngsi-ld:entity:E1
+
+
+
+02. PATCH E1, modifying both P1 and R1, with Content-Type application/merge-patch+json
+======================================================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+03. PATCH E1/P1, with Content-Type application/merge-patch+json
+===============================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+04. GET E1 - make sure step 02 and 03 worked
+============================================
+HTTP/1.1 200 OK
+Content-Length: 143
+Content-Type: application/json
+Date: REGEX(.*)
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+
+{
+    "P1": {
+        "type": "Property",
+        "value": 3
+    },
+    "R1": {
+        "object": "urn:ngsi-ld:entity:E2",
+        "type": "Relationship"
+    },
+    "id": "urn:ngsi-ld:entity:E1",
+    "type": "Device"
+}
+
+
+--TEARDOWN--
+brokerStop CB
+dbDrop CB
