+++ test/functionalTest/cases/0000_ngsild/ngsild_new_subattr-timestamps-patch-entity-attribute.test
@@ -1,0 +1,240 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+Sub-Attribute system attributes for PATCH /entities/*/attrs/*
+
+--SHELL-INIT--
+# Copied from ngsild_subattr-timestamps-patch-entity-attribute.test
+export BROKER=orionld
+dbInit CB
+brokerStart CB 0 IPv4 -experimental
+
+--SHELL--
+
+#
+# 01. Create an entity E1 with attributes A1 having a sub-property MD1
+# 02. See E1 in mongo - see createdAt and modifiedAt for the sub-property MD1
+# 03. PATCH Entity/attrs/A1, replacing MD1, and adding MD2
+# 04. See E1 in mongo - see createdAt and modifiedAt for the sub-properties
+# 05. Retrieve E1 with sysAttrs on
+#
+
+echo "01. Create an entity E1 with attributes A1 having a sub-property MD1"
+echo "===================================================================="
+payload='{
+  "id": "urn:ngsi-ld:entities:E1",
+  "type": "T",
+  "A1": {
+    "type": "Property",
+    "value": "A1",
+    "MD1": {
+      "type": "Property",
+      "value": 1
+    }
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities --payload "$payload"
+echo
+echo
+
+
+echo "02. See E1 in mongo - see createdAt and modifiedAt for the sub-property MD1"
+echo "==========================================================================="
+mongoCmd2 ftest "db.entities.findOne()"
+echo
+echo
+
+
+echo "03. PATCH Entity/attrs/A1, replacing MD1, and adding MD2"
+echo "========================================================"
+payload='{
+  "value": "A1-Step03",
+  "MD1": {
+    "type": "Property",
+    "value": 3
+  },
+  "MD2": {
+    "type": "Relationship",
+    "object": "urn:ngsi-ld:subattr:MD2"
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/A1 --payload "$payload" -X PATCH
+echo
+echo
+
+
+echo "04. See E1 in mongo - see createdAt and modifiedAt for the sub-properties"
+echo "========================================================================="
+mongoCmd2 ftest "db.entities.findOne()"
+echo
+echo
+
+
+echo "05. Retrieve E1 with sysAttrs on"
+echo "================================"
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1?options=sysAttrs
+echo
+echo
+
+
+--REGEXPECT--
+01. Create an entity E1 with attributes A1 having a sub-property MD1
+====================================================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1
+
+
+
+02. See E1 in mongo - see createdAt and modifiedAt for the sub-property MD1
+===========================================================================
+MongoDB shell version REGEX(.*)
+connecting to: REGEX(.*)
+MongoDB server version: REGEX(.*)
+{
+	"_id" : {
+		"id" : "urn:ngsi-ld:entities:E1",
+		"type" : "https://uri.etsi.org/ngsi-ld/default-context/T",
+		"servicePath" : "/"
+	},
+	"attrNames" : [
+		"https://uri.etsi.org/ngsi-ld/default-context/A1"
+	],
+	"attrs" : {
+		"https://uri=etsi=org/ngsi-ld/default-context/A1" : {
+			"type" : "Property",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
+			"value" : "A1",
+			"mdNames" : [
+				"https://uri.etsi.org/ngsi-ld/default-context/MD1"
+			],
+			"md" : {
+				"https://uri=etsi=org/ngsi-ld/default-context/MD1" : {
+					"type" : "Property",
+					"value" : 1,
+					"createdAt" : REGEX(.*),
+					"modifiedAt" : REGEX(.*)
+				}
+			}
+		}
+	},
+	"creDate" : REGEX(.*),
+	"modDate" : REGEX(.*),
+	"lastCorrelator" : ""
+}
+bye
+
+
+03. PATCH Entity/attrs/A1, replacing MD1, and adding MD2
+========================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+04. See E1 in mongo - see createdAt and modifiedAt for the sub-properties
+=========================================================================
+MongoDB shell version REGEX(.*)
+connecting to: REGEX(.*)
+MongoDB server version: REGEX(.*)
+{
+	"_id" : {
+		"id" : "urn:ngsi-ld:entities:E1",
+		"type" : "https://uri.etsi.org/ngsi-ld/default-context/T",
+		"servicePath" : "/"
+	},
+	"attrNames" : [
+		"https://uri.etsi.org/ngsi-ld/default-context/A1"
+	],
+	"attrs" : {
+		"https://uri=etsi=org/ngsi-ld/default-context/A1" : {
+			"type" : "Property",
+			"creDate" : REGEX(.*),
+			"modDate" : REGEX(.*),
+			"value" : "A1-Step03",
+			"mdNames" : [
+				"https://uri.etsi.org/ngsi-ld/default-context/MD1",
+				"https://uri.etsi.org/ngsi-ld/default-context/MD2"
+			],
+			"md" : {
+				"https://uri=etsi=org/ngsi-ld/default-context/MD1" : {
+					"type" : "Property",
+					"value" : 3,
+					"createdAt" : REGEX(.*),
+					"modifiedAt" : REGEX(.*)
+				},
+				"https://uri=etsi=org/ngsi-ld/default-context/MD2" : {
+					"type" : "Relationship",
+					"value" : "urn:ngsi-ld:subattr:MD2",
+					"createdAt" : REGEX(.*),
+					"modifiedAt" : REGEX(.*)
+				}
+			}
+		}
+	},
+	"creDate" : REGEX(.*),
+	"modDate" : REGEX(.*),
+	"lastCorrelator" : ""
+}
+bye
+
+
+05. Retrieve E1 with sysAttrs on
+================================
+HTTP/1.1 200 OK
+Content-Length: 505
+Content-Type: application/json
+Date: REGEX(.*)
+Link: <https://uri.etsi.org/ngsi-ld/v1/ngsi-ld-core-context.jsonld>; rel="http://www.w3.org/ns/json-ld#context"; type="application/ld+json"
+
+{
+    "A1": {
+        "MD1": {
+            "createdAt": "REGEX(.*)",
+            "modifiedAt": "REGEX(.*)",
+            "type": "Property",
+            "value": 3
+        },
+        "MD2": {
+            "createdAt": "REGEX(.*)",
+            "modifiedAt": "REGEX(.*)",
+            "object": "urn:ngsi-ld:subattr:MD2",
+            "type": "Relationship"
+        },
+        "createdAt": "REGEX(.*)",
+        "modifiedAt": "REGEX(.*)",
+        "type": "Property",
+        "value": "A1-Step03"
+    },
+    "createdAt": "REGEX(.*)",
+    "id": "urn:ngsi-ld:entities:E1",
+    "modifiedAt": "REGEX(.*)",
+    "type": "T"
+}
+
+
+--TEARDOWN--
+brokerStop CB
+dbDrop CB
