+++ test/functionalTest/cases/0000_troe/troe_new_patch_attribute.test
@@ -1,0 +1,197 @@
+# Copyright 2022 FIWARE Foundation e.V.
+#
+# This file is part of Orion-LD Context Broker.
+#
+# Orion-LD Context Broker is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# Orion-LD Context Broker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+# General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with Orion-LD Context Broker. If not, see http://www.gnu.org/licenses/.
+#
+# For those usages not covered by this license please contact with
+# orionld at fiware dot org
+
+# VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
+
+--NAME--
+Simple test of creation of an Entity followed by a PATCH of one of the attributes of the entity
+
+--SHELL-INIT--
+# Copied from troe_patch_attribute.test
+export BROKER=orionld
+dbInit CB
+pgInit $CB_DB_NAME
+brokerStart CB 0 IPv4 -troe -experimental
+
+--SHELL--
+
+#
+# 01. Create an entity E1 with attributes P and R - P with sub-attrs Sub-R and Sub-Int
+# 02. PATCH E1:P, replacing Sub-R and adding Sub-P + 6 geo-props, and changing the attribute value to 44
+# 03. See E1 in the temporal database
+# 04. See the attributes in the temporal database
+# 05. See the sub-attributes in the temporal database
+#
+
+echo "01. Create an entity E1 with attributes P and R - P with sub-attrs Sub-R and Sub-Int"
+echo "===================================================================================="
+payload='{
+  "id": "urn:ngsi-ld:entities:E1",
+  "type": "T",
+  "P": {
+    "type": "Property",
+    "value": "P of E1",
+    "observedAt": "2020-12-10T07:11:00.123Z",
+    "Sub-R": {
+      "type": "Relationship",
+      "object": "urn:ngsi-ld:rel:R",
+      "observedAt": "2020-12-10T07:12:00.123Z"
+    },
+    "Sub-Int": {
+      "type": "Property",
+      "value": 33,
+      "observedAt": "2020-12-10T07:12:00.128Z",
+      "unitCode": "pi-i"
+    }
+  },
+  "R": {
+    "type": "Relationship",
+    "object": "urn:ngsi-ld:relationships:R",
+    "observedAt": "2020-12-10T07:11:00.129Z"
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities --payload "$payload"
+echo
+echo
+
+
+echo "02. PATCH E1:P, replacing Sub-R and adding Sub-P + 6 geo-props, and changing the attribute value to 44"
+echo "======================================================================================================"
+payload='{
+  "value": 44,
+  "Sub-R": {
+    "type": "Relationship",
+    "object": "urn:ngsi-ld:rel:R2",
+    "observedAt": "2020-12-10T07:12:00.124Z"
+  },
+  "Sub-P": {
+    "type": "Property",
+    "value": "Sub-P"
+  },
+  "geo-point": {
+    "type": "GeoProperty",
+    "value": {
+      "type": "Point",
+      "coordinates": [ 1,2,3 ]
+    }
+  },
+  "geo-line-string": {
+    "type": "GeoProperty",
+    "value": {
+      "type": "LineString",
+      "coordinates":  [ [ 1,2 ], [3,4], [5,6] ]
+    }
+  },
+  "geo-multi-line-string": {
+    "type": "GeoProperty",
+    "value": {
+      "type": "MultiLineString",
+      "coordinates":  [ [ [ 1,2 ], [3,4], [5,6] ], [ [ 1,2 ], [3,4], [5,6] ] ]
+    }
+  },
+  "geo-polygon": {
+    "type": "GeoProperty",
+    "value": {
+      "type": "Polygon",
+      "coordinates": [[ [0,0], [4,0], [4,-2], [0,-2], [0,0] ]]
+    }
+  },
+  "geo-multi-polygon": {
+    "type": "GeoProperty",
+    "value": {
+      "type": "MultiPolygon",
+      "coordinates": [ [[ [0,0], [4,0], [4,-2], [0,-2], [0,0] ]], [[ [10,0], [14,0], [14,-2], [10,-2], [10,0] ]] ]
+    }
+  }
+}'
+orionCurl --url /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1/attrs/P -X PATCH --payload "$payload"
+echo
+echo
+
+
+echo "03. See E1 in the temporal database"
+echo "==================================="
+postgresCmd -sql "SELECT opMode,id,type,ts FROM entities"
+echo
+echo
+
+
+echo "04. See the attributes in the temporal database"
+echo "==============================================="
+postgresCmd -sql "SELECT opMode,id,valueType,entityId,subProperties,unitcode,datasetid,text,number,boolean,observedAt,ts FROM attributes"
+echo
+echo
+
+
+echo "05. See the sub-attributes in the temporal database"
+echo "==================================================="
+postgresCmd -sql "SELECT id,valueType,entityId,attrInstanceId,unitcode,text,number,boolean,observedAt FROM subAttributes"
+echo
+echo
+
+
+--REGEXPECT--
+01. Create an entity E1 with attributes P and R - P with sub-attrs Sub-R and Sub-Int
+====================================================================================
+HTTP/1.1 201 Created
+Content-Length: 0
+Date: REGEX(.*)
+Location: /ngsi-ld/v1/entities/urn:ngsi-ld:entities:E1
+
+
+
+02. PATCH E1:P, replacing Sub-R and adding Sub-P + 6 geo-props, and changing the attribute value to 44
+======================================================================================================
+HTTP/1.1 204 No Content
+Date: REGEX(.*)
+
+
+
+03. See E1 in the temporal database
+===================================
+opmode,id,type,ts
+Create,urn:ngsi-ld:entities:E1,https://uri.etsi.org/ngsi-ld/default-context/T,REGEX(202.*)
+
+
+04. See the attributes in the temporal database
+===============================================
+opmode,id,valuetype,entityid,subproperties,unitcode,datasetid,text,number,boolean,observedat,ts
+Create,https://uri.etsi.org/ngsi-ld/default-context/P,String,urn:ngsi-ld:entities:E1,t,,None,P of E1,,,2020-12-10 07:11:00.123,REGEX(202.*)
+Create,https://uri.etsi.org/ngsi-ld/default-context/R,Relationship,urn:ngsi-ld:entities:E1,f,,None,urn:ngsi-ld:relationships:R,,,2020-12-10 07:11:00.129,REGEX(202.*)
+Update,https://uri.etsi.org/ngsi-ld/default-context/P,Number,urn:ngsi-ld:entities:E1,t,,None,,44,,,REGEX(202.*)
+
+
+05. See the sub-attributes in the temporal database
+===================================================
+id,valuetype,entityid,attrinstanceid,unitcode,text,number,boolean,observedat
+https://uri.etsi.org/ngsi-ld/default-context/Sub-R,Relationship,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,urn:ngsi-ld:rel:R,,,2020-12-10 07:12:00.123
+https://uri.etsi.org/ngsi-ld/default-context/Sub-Int,Number,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),pi-i,,33,,2020-12-10 07:12:00.128
+https://uri.etsi.org/ngsi-ld/default-context/Sub-R,Relationship,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,urn:ngsi-ld:rel:R2,,,2020-12-10 07:12:00.124
+https://uri.etsi.org/ngsi-ld/default-context/Sub-P,String,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,Sub-P,,,
+https://uri.etsi.org/ngsi-ld/default-context/geo-point,GeoPoint,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,,,,
+https://uri.etsi.org/ngsi-ld/default-context/geo-line-string,GeoLineString,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,,,,
+https://uri.etsi.org/ngsi-ld/default-context/geo-multi-line-string,GeoMultiLineString,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,,,,
+https://uri.etsi.org/ngsi-ld/default-context/geo-polygon,GeoPolygon,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,,,,
+https://uri.etsi.org/ngsi-ld/default-context/geo-multi-polygon,GeoMultiPolygon,urn:ngsi-ld:entities:E1,urn:ngsi-ld:attribute:instance:REGEX(.*),,,,,
+
+
+--TEARDOWN--
+brokerStop CB
+dbDrop CB
